requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
loguru==0.7.2
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0